webapp-serverside-express-guestbook: Create example
Express still seems to be a popular server-side framework, so let’s use
it for the first server-side version of the guestbook example webapp. I
didn’t know any Express before, so this might not be perfectly idiomatic
(pull requests welcome); development was slightly annoying at times but
overall not too bad.

The initial code was generated by

    npx express-generator --view=hbs webapp-serverside-express-guestbook

but I changed a lot of it (e.g. using ES6 modules, simplifying www/bin
given that we always have to use localhost:8080, using Wikimedia code
style, etc.).

For CSRF protection, this uses a tiny hand-written middleware, because
apparently the formerly recommended package (csurf) is deprecated with
no suggested replacement. For the session, we just use express-session’s
default MemoryStore, which should be good enough for this example.

The HTML is meant to be relatively close to the client-side Vite
guestbook example, though without the <div id="app">, as that seems
unnecessary in this case.

This uses the latest m3api-oauth2, so that we can correctly handle
situations where the user doesn’t strictly follow the expected flow
(immediately login, then try to sign the guestbook) without having to
track the state of the session ourselves. I’ll have to fix this in the
other example later.