webapp-*: Sync and update CSS
I meant to update the CSS of the server-side example before publishing
it, but then forgot to do that. So let’s update the CSS of both examples
at the same time: the styles are still fairly minimal (as mentioned in
the root README), but having a bit more than the browser defaults is
still nice. The design is meant to match Codex as of version 1.21.1 (see
[1], [2], [3]), but IMHO importing Codex properly would complicate the
examples too much – I’d rather just copy+paste a handful of design
tokens from it.

[1]: https://doc.wikimedia.org/codex/v1.21.1/style-guide/typography.html
[2]: https://doc.wikimedia.org/codex/v1.21.1/design-tokens/font.html
[3]: https://doc.wikimedia.org/codex/v1.21.1/design-tokens/spacing.html