webapp-clientside-vite-guestbook: Use separate OAuth callback
While working on the server-side example, I became unhappy with how the
code there had to accommodate for the OAuth callback being on the root
path, which isn’t realistic for a server-side app; also, it was
unsatisfying how the code (in both examples) had to detect the OAuth
callback URL by checking for the `code` callback parameter (which isn’t
exactly an implementation detail – it’s part of the OAuth standard – but
still, it’s knowledge that shouldn’t be needed outside of m3api-oauth2).
Luckily, it turns out that Vite supports multi-page apps relatively
simply [1], so I requested a new OAuth consumer with a different
callback URL [2] and wired up the app to use that. (The server-side
example will be updated separately.) The code in session.js is obviously
based on the one in the server-side example.

I’m not sure if this is going to be the usual way of writing MediaWiki
client-side apps going forward, but for now I think it’s an improvement
over the previous version.

[1]: https://vite.dev/guide/build.html#multi-page-app
[2]: https://meta.wikimedia.org/wiki/Special:OAuthListConsumers/view/3176b30484225b3b594b2a3074b01aff